feat: improvements in styles and components
