+++ src/components/Card/VideoCard.component.jsx
@@ -1,26 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { makeStyles } from '@material-ui/core/styles';
 import Card from '@material-ui/core/Card';
 import CardActionArea from '@material-ui/core/CardActionArea';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
 import Typography from '@material-ui/core/Typography';
 import parse from 'html-react-parser';
-
-const useStyles = makeStyles({
-  root: {
-    maxWidth: 345,
-    height: 350,
-  },
-});
+import useStyles from './VideoCard.styles';
 
 export default function VideoCard(props) {
   const classes = useStyles();
   const { title, image, description } = props;
 
   return (
-    <Card className={classes.root}>
+    <Card className={classes.card_size}>
       <CardActionArea>
         <CardMedia
           component="img"
