+++ src/components/Card/VideoCard.styles.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles({
+  card_size: {
+    maxWidth: 345,
+    height: 350,
+  },
+});
+
+export default useStyles;
