+++ src/components/Layout/Header/Menu/Menu.component.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import MenuItem from '@material-ui/core/MenuItem';
+import Menu from '@material-ui/core/Menu';
+
+export default function MenuBar(props) {
+  const isMenuOpen = Boolean(props.anchorEl);
+  const renderMenu = (
+    <Menu
+      anchorEl={props.anchorEl}
+      anchorOrigin={{ vertical: 'top', horizontal: 'right' }}
+      id={props.menuId}
+      keepMounted
+      transformOrigin={{ vertical: 'top', horizontal: 'right' }}
+      open={isMenuOpen}
+      onClose={props.handleMenuClose}
+    >
+      <MenuItem onClick={props.handleMenuClose}>Profile</MenuItem>
+      <MenuItem onClick={props.handleMenuClose}>My account</MenuItem>
+    </Menu>
+  );
+
+  return renderMenu;
+}
