+++ src/components/Layout/Header/MenuMobile/index.js
@@ -1,0 +1,1 @@
+export { default } from './MenuMobile.component';
