--- src/components/Layout/Layout.styles.css
@@ -1,9 +1,0 @@
-.container {
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin-top: -3rem;
-}
