+++ src/components/Layout/Layout.styles.js
@@ -1,0 +1,11 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles({
+  container: {
+    display: 'flex',
+    flexGrow: 1,
+    flexDirection: 'column',
+  },
+});
+
+export default useStyles;
