+++ src/pages/Home/Home.page.jsx
@@ -1,29 +1,16 @@
 import React from 'react';
 import Grid from '@material-ui/core/Grid';
-import { makeStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import useYoutube from '../../utils/hooks/useYoutube';
 import VideoCard from '../../components/Card';
-
-const useStyles = makeStyles({
-  title: {
-    padding: 30,
-    textAlign: 'center',
-  },
-});
+import Error from '../../components/ErrorMessage';
+import useStyles from './Home.styles';
 
 function HomePage() {
-  const { videosList, errorMessage } = useYoutube();
+  const { data, errorMessage } = useYoutube();
   const classes = useStyles();
   if (errorMessage) {
-    return (
-      <div>
-        <h2>
-          Error:
-          {errorMessage}
-        </h2>
-      </div>
-    );
+    return <Error errorMessage={errorMessage} />;
   }
   return (
     <>
@@ -31,7 +18,7 @@ function HomePage() {
         Welcome to the challenge!
       </Typography>
       <Grid container spacing={4}>
-        {videosList.map((video) => (
+        {data.map((video) => (
           <Grid item xs={6} md={3} key={video.etag}>
             <VideoCard
               title={video.snippet.title}
