+++ src/pages/Home/Home.styles.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles({
+  title: {
+    padding: 30,
+    textAlign: 'center',
+  },
+});
+
+export default useStyles;
