+++ src/utils/hooks/useYoutube.js
@@ -4,7 +4,7 @@ const API_URL =
   'https://gist.githubusercontent.com/jparciga/1d4dd34fb06ba74237f8966e2e777ff5/raw/f3af25f1505deb67e2cc9ee625a633f24d8983ff/youtube-videos-mock.json';
 
 function useYoutube() {
-  const [videosList, setVideosList] = useState([]);
+  const [data, setData] = useState([]);
   const [errorMessage, setErrorMessage] = useState(null);
 
   useEffect(() => {
@@ -12,7 +12,7 @@ function useYoutube() {
       try {
         const response = await fetch(API_URL);
         const result = await response.json();
-        setVideosList(result.items);
+        setData(result.items);
       } catch (error) {
         setErrorMessage(error.message);
       }
@@ -21,7 +21,7 @@ function useYoutube() {
     getVideos();
   }, []);
 
-  return { videosList, errorMessage };
+  return { data, errorMessage };
 }
 
 export default useYoutube;
